Feature/add settings proxy class (#35)
* added downloader to download any stream
added interface for downloadables
added tests
updated HTTP classes

* added settings

* using new settings class

* changed com to org
