+++ src/main/java/org/octopus/Main.java
@@ -22,15 +22,28 @@
  * SOFTWARE.
  */
 
-package com.octopus.core;
+package org.octopus;
 
-import java.io.InputStream;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
 
-public interface Downloadable {
-    /**
-     * Returns the input stream for a file
-     * @return InputStream
-     * @throws Exception Exception about HTTP
-     */
-    InputStream getFileStream() throws Exception;
+import java.io.IOException;
+
+public class Main extends Application {
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws IOException {
+        Parent root = FXMLLoader.load(getClass().getResource("/scenes/octopus.fxml"));
+        primaryStage.setTitle("Octopus");
+        Scene scene = new Scene(root, 600, 400);
+        primaryStage.setScene(scene);
+        primaryStage.show();
+    }
 }
