+++ src/main/java/org/octopus/core/http/HTTPUtils.java
@@ -22,28 +22,21 @@
  * SOFTWARE.
  */
 
-package com.octopus;
+package org.octopus.core.http;
 
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
+import java.net.URL;
 
-import java.io.IOException;
-
-public class Main extends Application {
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage primaryStage) throws IOException {
-        Parent root = FXMLLoader.load(getClass().getResource("/scenes/octopus.fxml"));
-        primaryStage.setTitle("Octopus");
-        Scene scene = new Scene(root, 600, 400);
-        primaryStage.setScene(scene);
-        primaryStage.show();
+public class HTTPUtils {
+    /**
+     * Extract the filename from the URL(with extension)
+     *
+     * @param url URL that to be inspected
+     * @return filename
+     */
+    public static String extractFileNameFromURL(URL url) {
+        String filename;
+        String path = url.getPath();
+        filename = path.substring(path.lastIndexOf('/') + 1);
+        return filename;
     }
 }
